资源分类信息查看，资源表列表：【【数据资源管理】-【资产目录】-资源分类信息查看-后端接口】https://www.tapd.cn/32773388/prong/tasks/view/1132773388001017819 联调接口： 【【数据资源管理】-【资产目录】-资源表查询】 https://www.tapd.cn/32773388/prong/stories/view/1132773388001017729

+++ src/main/java/com/hashtech/web/result/BaseInfo.java
@@ -20,7 +20,7 @@ public class BaseInfo {
     /**
      * 数据总条数，自动识别
      */
-    private Long dataSize;
+    private Integer dataSize;
     /**
      * 数据来源
      */
